# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "swagger documentation.",
        "title": "Best Lviv api docs",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/v1/ai/run": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI"
                ],
                "summary": "Manually trigger the prediction run (admin)",
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    }
                }
            }
        },
...
        "/v1/demand-readings": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Stores a new demand data point for a delivery point + resource pair. Triggers background analysis.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI"
                ],
                "summary": "Record a demand reading",
                "parameters": [
                    {
                        "description": "Demand reading",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httprequest.RecordDemand"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpresponse.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "Data": {
                                            "$ref": "#/definitions/entity.DemandReading"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    }
                }
            }
        },
        "/v1/demand-readings/{point_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI"
                ],
                "summary": "Get demand readings for a delivery point",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Customer/delivery point ID",
                        "name": "point_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    }
                }
            }
        },
...
        "/v1/predictive-alerts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI"
                ],
                "summary": "List open predictive alerts",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    }
                }
            }
        },
        "/v1/predictive-alerts/{alert_id}/dismiss": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI"
                ],
                "summary": "Dismiss a predictive alert",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Alert ID",
                        "name": "alert_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    }
                }
            }
        },
        "/v1/predictive-alerts/{point_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI"
                ],
                "summary": "Get all alerts for a delivery point",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Delivery point ID",
                        "name": "point_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    }
                }
            }
        },
        "/v1/rebalancing-proposals/{proposal_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI"
                ],
                "summary": "Get a rebalancing proposal with its transfers",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Proposal ID",
                        "name": "proposal_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpresponse.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "Data": {
                                            "$ref": "#/definitions/entity.RebalancingProposal"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/rebalancing-proposals/{proposal_id}/approve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI"
                ],
                "summary": "Approve a rebalancing proposal (one-tap confirm)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Proposal ID",
                        "name": "proposal_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpresponse.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "Data": {
                                            "$ref": "#/definitions/entity.RebalancingProposal"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/rebalancing-proposals/{proposal_id}/dismiss": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI"
                ],
                "summary": "Dismiss a rebalancing proposal",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Proposal ID",
                        "name": "proposal_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.DemandReading": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "point_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "number"
                },
                "recorded_at": {
                    "type": "string"
                },
                "resource_id": {
                    "type": "integer"
                },
                "source": {
                    "$ref": "#/definitions/entity.DemandSource"
                }
            }
        },
        "entity.DemandSource": {
            "type": "string",
            "enum": [
                "manual",
                "sensor",
                "predicted"
            ],
            "x-enum-varnames": [
                "DemandSourceManual",
                "DemandSourceSensor",
                "DemandSourcePredicted"
            ]
        },
...
        "entity.ProposalStatus": {
            "type": "string",
            "enum": [
                "pending",
                "approved",
                "dismissed"
            ],
            "x-enum-varnames": [
                "ProposalStatusPending",
                "ProposalStatusApproved",
                "ProposalStatusDismissed"
            ]
        },
        "entity.RebalancingProposal": {
            "type": "object",
            "properties": {
                "confidence": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "resource_id": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/entity.ProposalStatus"
                },
                "target_point_id": {
                    "type": "integer"
                },
                "transfers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.RebalancingTransfer"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "urgency": {
                    "type": "string"
                }
            }
        },
        "entity.RebalancingTransfer": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "estimated_arrival_hours": {
                    "type": "number"
                },
                "from_warehouse_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "proposal_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "number"
                }
            }
        },
...
        "httprequest.RecordDemand": {
            "type": "object",
            "required": [
                "point_id",
                "quantity",
                "resource_id"
            ],
            "properties": {
                "point_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "number"
                },
                "recorded_at": {
                    "type": "string"
                },
                "resource_id": {
                    "type": "integer"
                },
                "source": {
                    "enum": [
                        "manual",
                        "sensor",
                        "predicted"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.DemandSource"
                        }
                    ]
                }
            }
        },
...
        "httpresponse.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "metadata": {
                    "$ref": "#/definitions/httpresponse.ResponseMetadata"
                }
            }
        },
        "httpresponse.ResponseMetadata": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token, \"Bearer \u003ctoken\u003e\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}